requirements.txt
annotated-types==0.6.0
anyio==4.3.0
debugpy==1.8.0
appdirs==1.4.4
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
confluent-kafka==2.8.0
contourpy==1.2.0
cryptography==44.0.0
cycler==0.12.1
Django==5.0
django-currentuser==0.6.1
django-summernote==0.8.20.0
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.3
Flask==3.0.2
fonttools==4.50.0
frozendict==2.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
lxml==5.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
modcall==0.1.0
motor==3.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
orjson==3.10.3
outcome==1.3.0.post0
owlready2==0.45
packaging==24.0
passlib==1.7.4
pillow==10.3.0
pluggy==1.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pyDatalog==0.17.4
Pygments==2.18.0
PyJWT==2.10.1
PyLD==2.0.4
pymongo==4.7.2
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rdflib==7.0.0
requests==2.31.0
rich==13.7.1
selenium==4.18.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.4.4
starlette==0.37.2
sympy==1.12
tomli==2.0.1
trio==0.25.0
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.11.0
tzdata==2023.4
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.20.0
watchfiles==0.21.0
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
